smile Recognition/requirements.txt
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.0.2
opencv-python-headless==4.11.0.86
pillow==11.2.1
requests==2.32.3
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
